django_part/requirements.txt
Django==5.0.2
python-dotenv==1.0.1
django-cors-headers==4.3.1
Pillow==10.2.0
google-generativeai==0.3.2